% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GenerateDocs}
\alias{GenerateDocs}
\title{GenerateDocs}
\usage{
GenerateDocs(nDocs, node, b, eta, delta, sigma_tau, support, netstat, timestat,
  base.data = NULL, backward = FALSE, base = FALSE, timeunit = 3600,
  tz = "America/New_York")
}
\arguments{
\item{nDocs}{number of documents to be generated}

\item{node}{vector of node id's (ID starting from 1)}

\item{b}{coefficients for recipients}

\item{eta}{coefficients for timestamps}

\item{delta}{tuning parameter for the number of recipients}

\item{sigma_tau}{variance parameter for the timestamps}

\item{support}{support of latent recipients}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("sender", "receiver", "timeofday", "dayofweek")}

\item{base.data}{edges before 384 hours that is used to calculate initial history of interactions}

\item{backward}{Logigal indicating whether we are generating backward samples (if FALSE -> forward)}

\item{base}{Logical indicating whether or not we are generating base edges (< 384)}

\item{timeunit}{hour (= 3600) or day (=3600*24) and so on}

\item{tz}{timezone such as EST and PST}
}
\value{
generated data including (author, recipients, timestamp, words)
}
\description{
Generate a collection of documents according to the generative process of IPTM
}
