% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Inference}
\alias{Inference}
\title{Inference}
\usage{
Inference(edge, node, sigma.Q, prior.b, prior.delta, prior.eta, prior.tau,
  Outer, Inner, netstat, timestat, initial = NULL, timeunit = 3600,
  tz = "America/New_York")
}
\arguments{
\item{edge}{list of tie data with 3 elements (1: author, 2: recipient, 3: timestamp in unix.time format)}

\item{node}{vector of node id's (ID starting from 1)}

\item{sigma.Q}{proposal distribution variance parameter}

\item{prior.b}{prior mean and covariance of b in multivariate Normal distribution}

\item{prior.delta}{prior mean and variance of delta in Normal distribution}

\item{prior.eta}{prior mean and covariance of eta in multivariate Normal distribution}

\item{prior.tau}{prior shape and scale parameter of sigma_tau in inverse-Gamma distribution}

\item{Outer}{size of outer iterations}

\item{Inner}{size of inner iteration for Metropolis-Hastings updates}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("sender", "receiver","timeofday", "dayofweek")}

\item{initial}{list of initial values user wants to assign including (delta, b, eta, cd, z, u, sigma_tau, proposal.var1, proposal.var2)}

\item{timeunit}{hour (= 3600) or day (=3600*24) and so on}

\item{tz}{timezone such as EST and PST}
}
\value{
MCMC output containing all parameter estimates
}
\description{
Iterate Markov Chain Monte Carlo (MCMC) algorithm for the interaction-partitioned topic model
}
